Make the frequency estimate less jumpy
The process noise parameter (Q) is too big with respect to the
measurement noise because I didn't scale it with the duration of the
sample size.  Additionally, I simplify several things by using the
variance instead of standard deviation.  Oftentimes, I was just
squaring the std to get back the variance.